caption/requirements.txt
fastapi
uvicorn[standard]
python-multipart # For file uploads
Pillow # For image handling
transformers[torch] # Or [tf] if using TensorFlow
torch # Or tensorflow
accelerate # Often needed by transformers
requests # For potential internal communication or testing
pytest # For tests
# requirements.txt for caption and object backends
prometheus-client
# For FastAPI auto-instrumentation (RECOMMENDED):
prometheus-fastapi-instrumentator